7. Create a custom filter
Create a custom filter that lets you decide on which columns you can search. This filter must be
two controls, first, a select control to decide the columns to search in, and the second control
must be an input text containing the string to search in the selected column. The first control
must be dynamic, this means if we need to add another column to the table the select will get